mainApp/requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.1
django-daterange-filter==1.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
PyMuPDF==1.16.10
pyTelegramBotAPI==3.6.6
pytz==2019.3
requests==2.7.0
six==1.13.0
sqlparse==0.3.0
style==1.1.0
telebot==0.0.3
update==0.0.1
whitenoise==5.0.1
